+++ engine/src/vulkan/rendertarget.cpp
@@ -44,8 +44,8 @@ Result Context::viewport(rhi::Extent2D extent)
     VkViewport viewport = {};
     viewport.x = 0;
     viewport.y = 0;
-    viewport.width = extent.width;
-    viewport.height = extent.height;
+    viewport.width = static_cast<float>(extent.width);
+    viewport.height = static_cast<float>(extent.height);
     viewport.minDepth = 0.f;
     viewport.maxDepth = 1.f;
